fix__btn_play_stop
